Move tests and assertions where they belong
