+++ tests/plugins/test_history.py
@@ -26,6 +26,7 @@ class HelloViewTest(HistoryWebTest):
         self.assertIn("history", capabilities)
 
 
+@skip_if_no_statsd
 class MetricsTest(HistoryWebTest):
     @classmethod
     def get_app_settings(cls, extras=None):
@@ -37,7 +38,6 @@ class MetricsTest(HistoryWebTest):
         )
         return settings
 
-    @skip_if_no_statsd
     def test_a_statsd_timer_is_used_for_history_if_configured(self):
         with mock.patch("kinto.plugins.statsd.StatsDService.timer") as mocked:
             self.app.put("/buckets/test", headers=self.headers)
